refactor: standardize entity-count getters on u32 (task 033 T1)
len()/doc_count() returned an inconsistent mix of usize/u32/u64. Settle on u32 for
index-total entity counts — the doc-ID space is u32 system-wide (vector_id ==
doc_id), so a doc/term/vector count can't exceed it: TermIndex::len (was usize),
VectorIndex::len, RerankStore::len, ImpactIndex::doc_count, ImpactsAccumulator::
doc_count (all were u64). Backing on-disk u64 fields are unchanged (IDF math reads
them directly); only the public getters narrow. Genuine byte-lengths
(MemoryFetch::len) stay usize. Drops the now-redundant as-u32 casts in the wasm
RrtIndex/RrviIndex len bindings; per-key cardinality and cost sums (Index::
term_count/query_cost/count_estimate) stay u64 (they can exceed the ID space).

+++ rust/examples/build_impacts.rs
@@ -422,7 +422,7 @@ fn merge(
     let term_count: u32 = term_i
         .try_into()
         .map_err(|_| io::Error::other("term count exceeds u32"))?;
-    if term_count as usize != rrt_idx.len() {
+    if term_count != rrt_idx.len() {
         return Err(io::Error::other(format!(
             "merged {term_count} terms but the dictionary has {}",
             rrt_idx.len()
